# swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "User Game API",
        "version": "1.1.0",
        "description": "Simple game api that store user account, biodata, and game histories"
    },
    "schemes": [
        "http",
        "https"
    ],
    "securityDefinitions": {
        "jwt": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header",
            "description": "Example: Bearer [JWT TOKEN]"
        }
    },
    "consumes": [
        "application/json",
        "multipart/form-data"
    ],
    "produces": [
        "application/json"
    ],
    "basePath": "/api/v1",
...
    "paths": {
...
        "/user_games": {
            "get": {
                "summary": "Get all user games",
                "operationId": "GetUserGames",
                "responses": {
                    "200": {
                        "description": "Returns a list of user games",
                        "schema": {
                            "$ref": "#/responses/getUserGamesResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/responses/generalErrorResponse"
                        }
                    }
                },
                "tags": [
                    "User Game"
                ]
            },
            "post": {
                "summary": "Create a new user game",
                "operationId": "CreateUserGames",
                "responses": {
                    "201": {
                        "description": "Returns user game that has been created",
                        "schema": {
                            "$ref": "#/responses/postUserGamesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/responses/badRequestResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/responses/generalErrorResponse"
                        }
                    }
                },
                "tags": [
                    "User Game"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Create user game",
                        "required": true,
                        "schema": {
                            "$ref": "#/requestBodies/postUserGamesRequest"
                        }
                    }
                ]
            }
        },
        "/user_game/{id}": {
            "get": {
                "summary": "Get user game by id",
                "operationId": "GetUserGame",
                "responses": {
                    "200": {
                        "description": "Returns a user game",
                        "schema": {
                            "$ref": "#/responses/getUserGameResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/responses/badRequestResponse"
                        }
                    },
                    "404": {
                        "description": "User game not found",
                        "schema": {
                            "$ref": "#/responses/generalErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/responses/generalErrorResponse"
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User game id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "tags": [
                    "User Game"
                ]
            },
            "patch": {
                "summary": "Update user game by id",
                "operationId": "UpdateUserGame",
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns before after user game field that has been updated",
                        "schema": {
                            "$ref": "#/responses/patchUserGameResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/responses/badRequestResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/responses/generalErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/responses/generalErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User game not found",
                        "schema": {
                            "$ref": "#/responses/generalErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/responses/generalErrorResponse"
                        }
                    }
                },
                "tags": [
                    "User Game"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User game id",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Update user game",
                        "required": true,
                        "schema": {
                            "$ref": "#/requestBodies/patchUserGameRequest"
                        }
                    }
                ]
            },
            "delete": {
                "summary": "Delete user game by id",
                "operationId": "DeleteUserGame",
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns user game that has been deleted",
                        "schema": {
                            "$ref": "#/responses/deleteUserGameResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/responses/badRequestResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/responses/generalErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/responses/generalErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User game not found",
                        "schema": {
                            "$ref": "#/responses/generalErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/responses/generalErrorResponse"
                        }
                    }
                },
                "tags": [
                    "User Game"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User game id",
                        "required": true,
                        "type": "integer"
                    }
                ]
            }
        },
...
        "/user_game/{userGameId}/biodata": {
            "get": {
                "summary": "Get user game biodata by user game id",
                "operationId": "GetUserGameBiodataByUserGameId",
                "responses": {
                    "200": {
                        "description": "Returns a user game biodata by user game id",
                        "schema": {
                            "$ref": "#/responses/getUserGameBiodataByUserGameIdResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/responses/badRequestResponse"
                        }
                    },
                    "404": {
                        "description": "User game biodata not found",
                        "schema": {
                            "$ref": "#/responses/generalErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/responses/generalErrorResponse"
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "userGameId",
                        "in": "path",
                        "description": "User game id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "tags": [
                    "User Game Biodata"
                ]
            }
        },
...
        "/user_game/{userGameId}/histories": {
            "get": {
                "summary": "Get user game history by user game id",
                "operationId": "GetUserGameHistoryByUserGameId",
                "responses": {
                    "200": {
                        "description": "Returns a user game history by user game id",
                        "schema": {
                            "$ref": "#/responses/getUserGameHistoryByUserGameIdResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/responses/badRequestResponse"
                        }
                    },
                    "404": {
                        "description": "User game history not found",
                        "schema": {
                            "$ref": "#/responses/generalErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/responses/generalErrorResponse"
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "userGameId",
                        "in": "path",
                        "description": "User game id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "tags": [
                    "User Game History"
                ]
            }
        }
    },
...
    "responses": {
...
        "getUserGamesResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "integer",
                    "format": "int32",
                    "default": 200
                },
                "message": {
                    "type": "string",
                    "format": "string",
                    "default": "OK"
                },
                "count": {
                    "type": "integer",
                    "format": "int32"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/userGameInclude"
                    }
                }
            }
        },
        "postUserGamesResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "integer",
                    "format": "int32",
                    "default": 201
                },
                "message": {
                    "type": "string",
                    "format": "string",
                    "default": "UserGame created successfully"
                },
                "data": {
                    "$ref": "#/definitions/userGame"
                }
            }
        },
        "getUserGameResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "integer",
                    "format": "int32",
                    "default": 200
                },
                "message": {
                    "type": "string",
                    "format": "string",
                    "default": "OK"
                },
                "data": {
                    "$ref": "#/definitions/userGameInclude"
                }
            }
        },
        "patchUserGameResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "integer",
                    "format": "int32",
                    "default": 200
                },
                "message": {
                    "type": "string",
                    "format": "string",
                    "default": "UserGame updated successfully"
                },
                "data": {
                    "type": "object",
                    "properties": {
                        "before": {
                            "$ref": "#/requestBodies/patchUserGameRequest"
                        },
                        "after": {
                            "$ref": "#/requestBodies/patchUserGameRequest"
                        }
                    }
                }
            }
        },
        "deleteUserGameResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "integer",
                    "format": "int32",
                    "default": 200
                },
                "message": {
                    "type": "string",
                    "format": "string",
                    "default": "UserGame deleted successfully"
                },
                "data": {
                    "$ref": "#/definitions/userGame"
                }
            }
        },
...
        "getUserGameBiodataByUserGameIdResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "integer",
                    "format": "int32",
                    "default": 200
                },
                "message": {
                    "type": "string",
                    "format": "string",
                    "default": "OK"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/userGameBiodata"
                    }
                }
            }
        },
...
        "getUserGameHistoryByUserGameIdResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "integer",
                    "format": "int32",
                    "default": 200
                },
                "message": {
                    "type": "string",
                    "format": "string",
                    "default": "OK"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/userGameHistory"
                    }
                }
            }
        },
        "generalErrorResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "integer",
                    "format": "int32"
                },
                "message": {
                    "type": "string",
                    "format": "string"
                }
            }
        },
        "badRequestResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "integer",
                    "format": "int32",
                    "default": 400
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "value": {
                                "type": "string",
                                "format": "string"
                            },
                            "msg": {
                                "type": "string",
                                "format": "string"
                            },
                            "param": {
                                "type": "string",
                                "format": "string"
                            },
                            "location": {
                                "type": "string",
                                "format": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "userGame": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "username": {
                    "type": "string",
                    "format": "string"
                },
                "password": {
                    "type": "string",
                    "format": "string"
                },
                "createdAt": {
                    "type": "string",
                    "format": "date-time"
                },
                "updatedAt": {
                    "type": "string",
                    "format": "date-time"
                }
            }
        },
        "userGameInclude": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int32"
                },
                "username": {
                    "type": "string",
                    "format": "string"
                },
                "password": {
                    "type": "string",
                    "format": "string"
                },
                "createdAt": {
                    "type": "string",
                    "format": "date-time"
                },
                "updatedAt": {
                    "type": "string",
                    "format": "date-time"
                },
                "UserGameBiodatum": {
                    "$ref": "#/definitions/userGameBiodata"
                },
                "UserGameHistories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/userGameHistory"
                    }
                }
            }
        },
        "userGameBiodata": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int32"
                },
                "email": {
                    "type": "string",
                    "format": "string"
                },
                "firstname": {
                    "type": "string",
                    "format": "string"
                },
                "lastname": {
                    "type": "string",
                    "format": "string"
                },
                "profilePicture": {
                    "type": "string",
                    "format": "string"
                },
                "country": {
                    "type": "string",
                    "format": "string"
                },
                "age": {
                    "type": "integer",
                    "format": "int32"
                },
                "createdAt": {
                    "type": "string",
                    "format": "date-time"
                },
                "updatedAt": {
                    "type": "string",
                    "format": "date-time"
                },
                "userGameId": {
                    "type": "integer",
                    "format": "int32"
                }
            }
        },
...
        "userGameHistory": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int32"
                },
                "title": {
                    "type": "string",
                    "format": "string"
                },
                "publisher": {
                    "type": "string",
                    "format": "string"
                },
                "cover": {
                    "type": "string",
                    "format": "string"
                },
                "lastPlayed": {
                    "type": "string",
                    "format": "date-time"
                },
                "score": {
                    "type": "number",
                    "format": "int32"
                },
                "createdAt": {
                    "type": "string",
                    "format": "date-time"
                },
                "updatedAt": {
                    "type": "string",
                    "format": "date-time"
                },
                "userGameId": {
                    "type": "integer",
                    "format": "int32"
                }
            }
        },
...
    }
}